% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{bestCViter}
\alias{bestCViter}
\title{Get the best interation in terms of CV error (log-likelihood)}
\usage{
bestCViter(object)
}
\arguments{
\item{object}{An object of class "xgbm"}
}
\description{
Get the best interation in terms of CV error (log-likelihood)
}
