% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{big.model.matrix}
\alias{big.model.matrix}
\title{Create a model frame when there are many predictors.}
\usage{
big.model.matrix(fo, data, n.cores)
}
\arguments{
\item{fo}{A formula.}

\item{data}{A data frame.}

\item{n.cores}{The number of cores to use in parallel processing.}
}
\description{
Create a model frame when there are many predictors.
}
\details{
The function breaks the formula's right had side into it's
  component part, and uses each to construct a model matrix without an
  intercept. So the returned object has no intercept, but \code{xgboost}
  ought to be able to cope.
}
