% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgbm_methods.R
\name{fitted.xgbm}
\alias{fitted.xgbm}
\title{Fitted values from an xgbm object}
\usage{
\method{fitted}{xgbm}(object, type = "cv", ...)
}
\arguments{
\item{object}{An object of class 'xgbm'.}

\item{type}{String giving the type of fitted values. Defaults to 'cv' and
the fitted values are the test set cross-validation predictions. The only
other option is 'model' in which case the fitted model, at the best
cross-validation iteration, is run on the data used to train it.}
}
\description{
Fitted values from an xgbm object
}
