% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonxgboost.R
\name{getSynonyms}
\alias{getSynonyms}
\title{Get similar strings in a vector of strings}
\usage{
getSynonyms(x, terms, maxdist = 0.1)
}
\arguments{
\item{x}{The vector of strings to be searched.}

\item{terms}{Vetor of strings to be approximately matched.}

\item{maxdist}{Passed to \code{agrepl} as \code{max.distance}.}
}
\description{
Get similar strings in a vector of strings
}
