% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgbm_methods.R
\name{ggplot.xgbm}
\alias{ggplot.xgbm}
\title{Partial dependence plots for xgbm}
\usage{
\method{ggplot}{xgbm}(
  data,
  mapping = NULL,
  i.var = 1,
  return.grid = FALSE,
  type = "auto",
  color = "blue",
  ...
)
}
\arguments{
\item{data}{An object of class 'xgbm'.}

\item{mapping}{Not used.}

\item{i.var}{A number giving the predictor variable to plot (the function
assumes there is no intercept), or a string giving the variable name.}

\item{return.grid}{Whether to return the grid of data points and predicted
values. Defaults to \code{return.grid = FALSE} and the plot is returned.}

\item{type}{String with value 'auto' (the default), 'regression' or 'classification'.
Passed through to \code{pdp::partial}.}
}
\description{
Partial dependence plots for xgbm
}
