% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhdm-package.R
\docType{package}
\name{mhdm-package}
\alias{mhdm-package}
\title{Gradient Boosting wrappers for xgboost}
\description{
Wrappers for xgboost functions, in the hope of making that package more
friendly to statisticians, especially those used to using the gbm package.
Also includes helper functions for data mining of medical history or
adverse event data.
}
