% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgbm_methods.R
\name{predict.xgbm}
\alias{predict.xgbm}
\title{Get predictions from an xgbm object}
\usage{
\method{predict}{xgbm}(x, newdata = NULL, n.trees = NULL, type = "link", n.cores = NULL)
}
\arguments{
\item{x}{A object of class 'xgbm'.}

\item{newdata}{A data frame. If not specified, the original training data is used.}

\item{n.trees}{The number of trees to use. If not specified, the number that
optimizes cross-validation error is used.}

\item{type}{String specifying either "link" or "response". Defaults to
\code{type = "link"} for compatibility with \code{gbm}, but will seldom
be desired.}
}
\description{
Get predictions from an xgbm object
}
