% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ri.R
\name{relative.influence}
\alias{relative.influence}
\title{Relative influence of predictors in an xgbm model}
\usage{
relative.influence(x)
}
\arguments{
\item{x}{An object of class 'xgbm'.}
}
\description{
Relative influence of predictors in an xgbm model
}
\details{
Relative influence is referred to as relative importance in the
  xgboost documentation, but I prefer 'influence' because it keeps the idea
  distinct from Breiman's variable importance. Also, it should be computed
  as the sum over all splits on a variable (according to the xgboost
  documentation), but isn't when there are factors in the model. So here
  we add the values associated with factors back together. Finally,
  the documentation for \code{xgb.importance} (for which this is a
  wrapper) states that you need to give in an integer vector of tree numbers
  indexed at 0 (ffs). This function takes care of it all by assuming you
  used cross-validation.
}
