% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonxgboost.R
\name{rrci}
\alias{rrci}
\title{Compute relative risk with approximate confidence intervals}
\usage{
rrci(x, nx, y, ny, alpha = 0.05, method = "Jeffreys")
}
\arguments{
\item{x}{The number of success in the first group}

\item{nx}{The number of observations in the first group}

\item{y}{The number of success in the second group}

\item{ny}{The number of observations in the second group}

\item{alpha}{Determines the confidence level. Defaults to \code{alpha=c(0.5, 0.1)}
and both 90\% and 50\% intervals are produced.}

\item{method}{Character string giving the method. Allowable options are
\code{method="MLE"} and \code{method="Jeffreys"}.(the default). It isn't case-sensitive}
}
\description{
Compute relative risk with approximate confidence intervals
}
\details{
If \code{method="Jeffreys"} (the default), the function adds 1/2 to
  the numerator and 1 to the denominator when estimating the
  proporions of successes in each group, thus avoiding infinite values. Then,
  estimate the standard error of the log relative risks, compute the confidence
  intervals, and transform back to the original scale.
}
