% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{stringsAsFactors}
\alias{stringsAsFactors}
\title{Deal with fallout from R 4.0 putting stringsAsFactors = FALSE}
\usage{
stringsAsFactors(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A data frame in which all character columns have been turned into factors.
}
\description{
Deal with fallout from R 4.0 putting stringsAsFactors = FALSE
}
