% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nonxgboost.R
\name{wideData}
\alias{wideData}
\title{Extract a wide form of the adverse event, medical history, conmeds, or other dataset}
\usage{
wideData(
  data,
  dm,
  subject = "subject",
  term = "PT",
  arm = "arm",
  arm.col = FALSE,
  subject.col = TRUE,
  drop = 0
)
}
\arguments{
\item{data}{The adverse event dataset}

\item{dm}{The demography dataset}

\item{subject}{The unique subject identifier. Defaults to \code{id="usubjid"}}

\item{term}{The name of the column identifying AE terms. Defaults to \code{term="aeterm"}}

\item{arm}{The name of the treatment group column. Defaults to \code{trt="arm"}}

\item{arm.col}{Whether to include the treatment arm as a column in the output. Defaults to \code{arm.col=FALSE}}

\item{subject.col}{Whether to include the subject identifier as a column in the output (the rownames get the subject identifiers). Defaults to \code{id.col=TRUE}}

\item{drop}{Drop columns in which fewer than \code{drop} of cases are 1. Defaults to \code{drop=0}}
}
\description{
Extract a wide form of the adverse event, medical history, conmeds, or other dataset
}
