% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgbm-package.R
\docType{package}
\name{xgbm-package}
\alias{xgbm-package}
\title{Gradient Boosting wrappers for xgboost}
\description{
Wrappers for xgboost functions, in the hope of making that package more
friendly to statisticians, especially those used to using the gbm package.
}
