% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{xgbm.perf}
\alias{xgbm.perf}
\title{Plot the training and testing set error over iterations}
\usage{
xgbm.perf(object, plot.it = TRUE)
}
\arguments{
\item{object}{An object of class 'xgbm'.}

\item{plot.it}{Whether to plot, defaulting to \code{plot.it = TRUE}.
If \code{plot.it = FALSE}, the best number of trees, as determined by
cross validation, is returned.}
}
\description{
Plot the training and testing set error over iterations
}
